refactor server functions with shared config and validation
